Web-Interface/Backend/requirements.txt

fastapi
uvicorn
pyserial
websockets
python-multipart
